[Oztechan/CCC#1952] Move single used mappers into their modules from network and database
