+++ client/datasource/currency/src/commonMain/kotlin/com/oztechan/ccc/client/datasource/currency/mapper/CurrencyMapper.kt
@@ -2,12 +2,13 @@
  * Copyright (c) 2021 Mustafa Ozhan. All rights reserved.
  */
 
-package com.oztechan.ccc.common.core.database.mapper
+package com.oztechan.ccc.client.datasource.currency.mapper
 
+import com.oztechan.ccc.common.core.database.mapper.toBoolean
 import com.oztechan.ccc.common.core.database.sql.Currency as CurrencyDBModel
 import com.oztechan.ccc.common.core.model.Currency as CurrencyModel
 
-fun CurrencyDBModel.toCurrencyModel() = CurrencyModel(
+internal fun CurrencyDBModel.toCurrencyModel() = CurrencyModel(
     code = code,
     name = name,
     symbol = symbol,
