+++ client/datasource/watcher/src/commonMain/kotlin/com/oztechan/ccc/client/datasource/watcher/mapper/WatcherMapper.kt
@@ -1,9 +1,10 @@
-package com.oztechan.ccc.common.core.database.mapper
+package com.oztechan.ccc.client.datasource.watcher.mapper
 
+import com.oztechan.ccc.common.core.database.mapper.toBoolean
 import com.oztechan.ccc.common.core.database.sql.Watcher as WatcherDBModel
 import com.oztechan.ccc.common.core.model.Watcher as WatcherModel
 
-fun WatcherDBModel.toWatcherModel() = WatcherModel(
+internal fun WatcherDBModel.toWatcherModel() = WatcherModel(
     id = id,
     base = base,
     target = target,
