+++ common/core/database/src/commonMain/kotlin/com/oztechan/ccc/common/core/database/di/CommonCoreDatabaseModule.kt
@@ -7,11 +7,11 @@ import org.koin.dsl.module
 private const val DATABASE_NAME = "application_database.sqlite"
 
 val commonCoreDatabaseModule = module {
+    single { provideDatabase(DATABASE_NAME) }
+
     single { get<CurrencyConverterCalculatorDatabase>().currencyQueries }
     single { get<CurrencyConverterCalculatorDatabase>().conversionQueries }
     single { get<CurrencyConverterCalculatorDatabase>().watcherQueries }
-
-    single { provideDatabase(DATABASE_NAME) }
 }
 
 expect fun Scope.provideDatabase(databaseName: String): CurrencyConverterCalculatorDatabase
