+++ common/core/database/src/commonMain/kotlin/com/oztechan/ccc/common/core/database/mapper/LongMapper.kt
@@ -1,6 +1,6 @@
 package com.oztechan.ccc.common.core.database.mapper
 
-internal fun Long.toBoolean() = when (this) {
+fun Long.toBoolean() = when (this) {
     1L -> true
     0L -> false
     else -> error("Value can not be boolean")
