+++ common/datasource/conversion/src/commonMain/kotlin/com/oztechan/ccc/common/datasource/conversion/mapper/ConversionMapper.kt
@@ -1,9 +1,9 @@
-package com.oztechan.ccc.common.core.database.mapper
+package com.oztechan.ccc.common.datasource.conversion.mapper
 
 import com.oztechan.ccc.common.core.model.Conversion
 import com.oztechan.ccc.common.core.database.sql.Conversion as ConversionDBModel
 
-fun ConversionDBModel.toConversionModel() = Conversion(
+internal fun ConversionDBModel.toConversionModel() = Conversion(
     base, date, AED, AFN, ALL, AMD, ANG, AOA, ARS, AUD, AWG, AZN, BAM, BBD, BDT,
     BGN, BHD, BIF, BMD, BND, BOB, BRL, BSD, BTC, BTN, BWP, BYN, BZD, CAD, CDF, CHF,
     CLF, CLP, CNH, CNY, COP, CRC, CUC, CUP, CVE, CZK, DJF, DKK, DOP, DZD, EGP, ERN,
